+++ include/nf_media.h
@@ -39,7 +39,7 @@ extern "C" {
 ///
 /// @param file File path.
 /// @param slot Slot number (0 - 15).
-void NF_LoadBMP(const char *file, u8 slot);
+void NF_LoadBMP(const char *file, u32 slot);
 
 /// @}
